requirements.txt
Flask==2.2.2
Flask-Cors==4.0.0
openai==1.45.1
python-dotenv==1.0.1
spacy==3.8.2
keybert==0.8.5
pymongo==4.6.0
fuzzywuzzy==0.18.0
numpy==1.26.4
scipy==1.10.1
torch==2.0.1
h5py==3.9.0



#source RecomSys/bin/activate
#pip install -r requirements.txt
#python -m spacy download en_core_web_sm 
#pip install pymongo